tests/on_target/requirements.txt
pytest
pyserial
termcolor
pyusb
imgtool
requests
memfault-cli
ppk2-api
pandas
plotly
